refactor: pages file structure
