+++ src/components/index.js
@@ -7,4 +7,5 @@ export { default as Subtitle } from './Subtitle'
 export { default as SubtitleAlternate } from './SubtitleAlternate'
 export { default as Image } from './image'
 export { default as SEO } from './seo'
+export { default as Success } from './success'
 export { default as ContentWrapper, TextContentWrapper } from './contentWrapper'
